convert unit image to use wolfi-base
it still uses concourse/golang-builder as its base, which is now based
on wolfi-base.

The one thing I haven't tested and am worried about is if the Puppeteer
tests will work on not. I'm not sure if the chromium we install will
work or not. Sadly that's the only "chrome" package in the wolfi
repository. Something to watch out for though. Might have to figure out
how to manually install chrome in the container which might not be very
straight-forward. Fingers crossed that it works!

Signed-off-by: Taylor Silva <dev@taydev.net>