adding openssl to hopefully resolve SSL errors during package installs
Every so often the image fails to build because the `apk add` command
gets this error:

```
SSL routines::unexpected eof while reading:ssl/record/rec_layer_s3.c
```

Not 100% sure this will fix the issue. There's usually another error
that comes right after that that'll look like this:

```
ERROR: iproute2-6.15.0-r0: Permission denied
```

But I'm guessing that's just a result of the SSL error as it happens for
different packages.

Searching the SSL error returns a bunch of results of folks saying to
update openssl, so that's what I'm attempting here.

Signed-off-by: Taylor Silva <dev@taydev.net>