break-up unit test tasks into separate jobs
lately it feels like one of the unit test tasks will flake which
requires re-running the entire job, which is a waste if the rest passed.
I want the flexibility to re-run just the unit tests that flaked.

As far as I can tell, none of the unit tests require the web part being
compiled, so the whole `built-concourse` artifact has been redundant.

Signed-off-by: Taylor Silva <dev@taydev.net>